Mac M2 benchmarks (#175)
Added Mac M2 benchmarks. Limit to 4 threads (performance cores). Adding
efficiency cores slows performance.

+++ samples/sample_shapes.cpp
@@ -715,8 +715,8 @@ public:
 			return true;
 		}
 
-		int indexA = reinterpret_cast<intptr_t>(userDataA);
-		int indexB = reinterpret_cast<intptr_t>(userDataB);
+		int indexA = static_cast<int>(reinterpret_cast<intptr_t>(userDataA));
+		int indexB = static_cast<int>(reinterpret_cast<intptr_t>(userDataB));
 
 		return ((indexA & 1) + (indexB & 1)) != 1;
 	}
